pyproject.toml
[project]
name = "single-photon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.21",
    "imageio>=2.37.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pylint>=4.0.2",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "streamlit>=1.51.0",
    "torchviz>=0.0.3",
    "visionsim>=0.1.0a1",
]
